Move UNIX and WINDOWS platform defines to separate header
Moving platform #defines to a header file so they can be
reused by other headers. This fixes build issues with
ofxhUtilities.cpp on Windows where a file that includes
ofxhBinary.h and ofxhUtilities.h will be able to call
utf8_to_utf16(), but will fail to link. This happens
because ofxhUtilities.o doesn't have the utf8_to_utf16()
implementation because WINDOWS was not defined. This
issue was discovered while doing a debug build of
openfx-misc on Windows. The DebugProxy plugin failed
to link because of this issue.

This change fixes the linking issue by making sure that
WINDOWS is defined properly when ofxhUtilities.h is included
and ensures that it is defined when ofxhUtilities.cpp is
built.